+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 
 ## Version Notes
+### 2.0.2
+Fix Web Authentication flow on Android returning Credentials object with idToken and accessToken mixed around
+Fix JSON serialization for Credentials, DatabaseUser, and UserInfo objects 
+
 ### 2.0.1
 Fix publish